requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0

# Data processing
opencv-python>=4.7.0
albumentations>=1.3.0
scikit-image>=0.20.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# DINOv3 backbone
# Note: DINOv3 should be installed separately from:
# https://github.com/facebookresearch/dinov3
# Clone the repo and install: pip install -e .
